fix method parse_tags
